+++ schema/postgres/test.sql
@@ -12,3 +12,4 @@ SET search_path TO public;
 \i /go/src/github.com/pirsch-analytics/pirsch/schema/postgres/v1.9.0.sql
 \i /go/src/github.com/pirsch-analytics/pirsch/schema/postgres/v1.9.1.sql
 \i /go/src/github.com/pirsch-analytics/pirsch/schema/postgres/v1.9.3.sql
+\i /go/src/github.com/pirsch-analytics/pirsch/schema/postgres/v1.10.0.sql